Wrote a function that converts integer to binary

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -7,7 +7,7 @@
 void print_binary(unsigned long int n)
 {
 	int i = 0, num;
-	unsigned int size = n;
+	unsigned long int size = n;
 
 	if (size == 0)
 	{
